requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.0
xgboost>=2.0
scipy>=1.11
joblib>=1.3
matplotlib>=3.7
requests>=2.31
python-dotenv>=1.0
ufcscraper>=0.2
